requirements.txt
# Core Airflow
apache-airflow==2.10.5

# Database
psycopg2-binary==2.9.9

# Web server
gunicorn==23.0.0

# Data processing
pandas==2.2.3
numpy==2.2.4

# HTTP requests
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0

# Configuration
python-dotenv==1.1.0

# API météo
openmeteo-requests==1.3.0

# Contraintes Airflow (important pour Python 3.12)
--constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.10.5/constraints-3.12.txt